fix(testing): fixes for travis test runs
