+++ package.json
@@ -15,7 +15,7 @@
     "old:test:cov": "jest --coverage && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
     "test": "npm run test:mocha",
     "tmp:test:cov": "npm run test:instrument && npm run test:mocha-cov && npm run coverage",
-    "test:cov": "npm run test:instrument && npm run test:mocha-cov && npm run test:coverage",
+    "test:cov": "npm run test:instrument && npm run test:mocha-cov && npm run coverage",
     "test:instrument": "babel-istanbul instrument src --output cov --no-compact --complete-copy --es-modules",
     "test:mocha": "electron-mocha ./tests/**/*.spec.js* --renderer --compilers js:babel-core/register --timeout 10000",
     "tmp:test:mocha-cov": "npm run test:mocha -- -r ./tests/coverage.js -R ./tests/istanbul-reporter.js",
