+++ tests/coverage.js
@@ -4,9 +4,6 @@ const originalRequire = Module.prototype.require;
 
 Module.prototype.require = function fancyCoverageRequireHack(moduleName, ...args) {
   try {
-    if (/src\//.test(moduleName)) {
-      console.log('[replacing] coverage file:', moduleName.replace('src/','cov/'));
-    }
     return originalRequire.call(this, moduleName.replace('src/', 'cov/'), ...args);
   }
   catch (e) {
