+++ tests/test.spec.jsx
@@ -38,19 +38,6 @@ describe('ScaleText', () => {
     windowHeight = window.outerHeight;
   });
 
-  after(function() {
-    if (window.__coverage__) {
-      console.log('Found coverage report, writing to coverage/coverage.json');
-      var path = require('path');
-      var fs = require('fs');
-      var coverageDir = path.resolve(process.cwd(), 'coverage');
-      if (!fs.existsSync()) {
-        fs.mkdirSync(coverageDir);
-      }
-      fs.writeFileSync(path.resolve(coverageDir, 'coverage.json'), JSON.stringify(window.__coverage__));
-    }
-  });
-
   afterEach((done) => {
     resetWindowSize();
     setTimeout(done, 500);
